% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs.R
\name{get_crs}
\alias{get_crs}
\title{get the coordinate reference system (CRS) for simple feature object}
\usage{
get_crs(sf_in_memory)
}
\arguments{
\item{sf_in_memory}{}
}
\value{
tibble of sf object names (sf) and their Wkt project information (WKt)
}
\description{
get the coordinate reference system (CRS) for simple feature object
}
\examples{
library(sfdep)
test_sf_object <- guerry # a built in dataset in the sfdep package
library(sf)
demo(nc, ask = FALSE, echo = FALSE)
see_sf() # should be test_sf_object and nc
sf_in_memory <- see_sf()
get_crs(sf_in_memory)
sf_in_memory |> get_crs()
}
