% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inla.R
\name{get_inla}
\alias{get_inla}
\title{get character vector of INLA objects in memory}
\usage{
get_inla()
}
\value{
a character vector of INLA objects in memory
}
\description{
get character vector of INLA objects in memory
}
\examples{
get_inla()
get_inla() |> mget()
test_inla_model <- inla(rnorm)
}
