% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{function for "Not In", the complement of \%in\%, like dplyr::anti_join but for vectors instead of data frame columns}
\usage{
... \%nin\% NA
}
\value{
a filtered vector
}
\description{
function for "Not In", the complement of \%in\%, like dplyr::anti_join but for vectors instead of data frame columns
}
\examples{
vowels <- c('A', 'E', 'I', 'O', 'U')
(consonants <- LETTERS \%nin\% vowels)

warpbreaks # has tension values of L, M, and H
tension_to_drop <- c('L', 'M')
warpbreaks |> dplyr::filter(tension \%nin\% tension_to_drop)
}
