% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LockeR-package.R
\docType{package}
\name{LockeR-package}
\alias{LockeR}
\alias{LockeR-package}
\title{LockeR: collection of R functions Dexter Locke uses all of the time}
\description{
Its where I keep my valuables. Groups utilities commonly used by the author. This is the worst description ever - sorry.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DHLocke/LockeR}
  \item Report bugs at \url{https://github.com/DHLocke/LockeR/issues}
}

}
\author{
\strong{Maintainer}: Dexter Locke \email{dexter.locke@gmail.com} (\href{https://orcid.org/0000-0003-2704-9720}{ORCID})

}
\keyword{internal}
