% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone_it_in.R
\name{phone_it_in}
\alias{phone_it_in}
\title{clean up US phone numbers}
\usage{
phone_it_in(phone, invalid = NA)
}
\arguments{
\item{phone}{string containing phone numbers to be standardized}

\item{invalid}{identifies candidates that are possibly not phone numbers at all}
}
\value{
phone numbers in a ###-#### or ###-###-#### format
}
\description{
Standardize 7 and 10 digit numbers to have consistent formatting without parentheses, with dashes between the 3rd and 4th digits and between the 6th and 7th digits when an area code is provided. Numbers not 7 or 10 digits in length are assigned NA by default, but can be assigned a chosen value with the "invalid" argument.
}
\examples{
# clean up US phone numbers
# based on
# https://stackoverflow.com/questions/34616466/formatting-phone-numbers-in-r
phone <- c("(123)-456-7890", "1234567890", "456890", "456-7890")
phone_it_in(phone)
phone_it_in(phone, invalid = 'Not a phone number')
}
