% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_within.R
\name{scale_y_reordered}
\alias{scale_y_reordered}
\title{Reorder bars within facets}
\usage{
scale_y_reordered(..., sep = "___")
}
\arguments{
\item{sep}{}
}
\value{
ggplot graph with elements ordered within a facet
}
\description{
Reorder bars within facets
}
\examples{
i <- data.frame(
nbr =c(15.18 ,11.53 ,13.37 ,9.2, 10.9, 12.23 ,9.53, 9.81, 7.86, 12.79,
22.03 ,17.64 ,18.1, 16.78 ,17.53 ,16.97 ,17.76 ,18.35 ,12.82 ,20.91,
22.09 ,19.18 ,17.54 ,18.45 ,19.83 ,16.99 ,19.69 ,19.45 ,13.07 ,21.41,
12.13 ,9.76, 10.79 ,10.74 ,12.43 ,9.65, 12.18 ,11.63 ,6.74, 12.31,
17.5, 14.75 ,15.2, 13.89 ,15.24 ,17.43 ,15.22 ,14.04,9.49, 15.86,
8.09, 5.86, 6.68, 7.34, 8.01, 6.35, 8.4, 7.4, 3.88, 6.92 ),
x2=rep(c("a",  "b",   "c",  "d", "e", "f", "g", "h", "i", "j"),6),
s = c(rep(c(rep(c("3"),10),
rep(c("4"),10),
rep(c("5"),10),
rep(c("6"),10),
rep(c("7"),10),
rep(c("8"),10)),1)))
ii <- i[order(i$s, i$nbr ), ]
sn <- factor(x = 1:60, labels = ii$x2)
ii$sn <- sn
ggplot(data = ii, aes(x = sn , y = nbr)) +
facet_wrap(. ~ s, ncol=2,scale='free_x')+   geom_bar(stat = 'identity')+
theme(axis.text.x=element_text(angle=90,hjust=1,vjust=.5,colour='gray50'))

ggplot(ii, aes(reorder_within(sn, nbr, s), nbr)) +
geom_bar(stat = 'identity') +
scale_x_reordered() +
facet_wrap(.~ s, ncol=2,scales = "free_x") +
theme(axis.text.x=element_text(angle=90,hjust=1,vjust=.5,colour='gray50'))

}
