% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_dirs.R
\name{set_up_dirs}
\alias{set_up_dirs}
\title{set up directories}
\usage{
set_up_dirs(proj_root, folders)
}
\arguments{
\item{proj_root}{the project's root folder, this is where the new folders will be created.}

\item{folders}{a vector of strings contianing the names of the new folders to make}
}
\description{
This creates commonly-used directories quickly, instead of manually making each folder, one at a time. If a directory with that name already exists in the chosen location, a new one is NOT created.
}
\examples{
(proj_root <- file.path(getwd()))
(folders <-
  c(
    'figures'
  , 'scripts'
  , 'input_data'
  , 'output_tables'
  , 'zipped_inputs'
  , 'saved_sessions'
  ))

  set_up_dirs(proj_root, folders)
}
