% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_replace_many.R
\name{str_replace_many}
\alias{str_replace_many}
\title{str_replace_many, like str_replace_all but accepts many from and to replacements at once
thanks Josiah Parry @josiahparry
originally posted here https://gist.github.com/JosiahParry/7d953f4af6848e7012566f7abfc63473}
\usage{
str_replace_many(string, ...)
}
\arguments{
\item{string}{the vector or data frame column to apply find/replacements upon}
}
\description{
This function does what stringr::str_replace_all does, but allows for multiple replacements at oncer
}
\examples{
str_replace_many("del hello wor!d", "del " = "", "!" = "l")

}
