% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_in_parens.R
\name{text_in_parens}
\alias{text_in_parens}
\title{extract text from within parentheses}
\usage{
text_in_parens(string_w_parens)
}
\arguments{
\item{string_w_parens}{string containing parentheses with text to be exracted}
}
\value{
the text between two parentheses
}
\description{
this finds and extracts the string between "(" and ")".
}
\examples{
# extract text from within parentheses
# based on
# https://community.rstudio.com/t/extract-text-between-brakets/43448/6
(letters_df <- data.frame(LETTERS, with_parens = paste0('(', LETTERS, ')')))
letters_df |> dplyr::mutate(full_circle = text_in_parens(with_parens))
}
