% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_cel.R
\name{to_cel}
\alias{to_cel}
\title{Title convert Fahrenheit to Celsius}
\usage{
to_cel(fahr)
}
\arguments{
\item{fahr}{degrees Fahrenheit}
}
\value{
the degrees Celsius
}
\description{
Title convert Fahrenheit to Celsius
}
\examples{
to_cel(32)   # 0  works!
to_cel(212)  # 100 works!
to_cel(-40)  # -40, success!
}
