% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_farh.R
\name{to_fahr}
\alias{to_fahr}
\title{Title convert Celsius to Fahrenheit}
\usage{
to_fahr(celsius)
}
\arguments{
\item{celsius}{degrees Celsius}
}
\value{
the degrees Fahrenheit
}
\description{
Title convert Celsius to Fahrenheit
}
\examples{
to_fahr(0)   # 32  works!
to_fahr(100) # 212 works!
to_fahr(-40)  # -40, success!
}
