% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs.R
\name{get_crs}
\alias{get_crs}
\title{get the coordinate reference system (CRS) for simple feature objects in memory}
\usage{
get_crs(sf_or_all = "sf")
}
\value{
tibble of sf object names (sf) and their epsg and Wkt projection information (WKt)
}
\description{
get the coordinate reference system (CRS) for simple feature objects in memory
}
\examples{
library(sf)
library(spData) # package of spatial data, see also 'spDataLarge'
data(alaska)
data(nz)
data(urban_agglomerations)
see_sf()
get_crs()
}
