% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crs.R
\name{get_crs}
\alias{get_crs}
\title{get the coordinate reference system (CRS) for simple feature object}
\usage{
get_crs(epsg = TRUE)
}
\arguments{
\item{sf_in_memory}{}
}
\value{
tibble of sf object names (sf) and their Wkt project information (WKt)
}
\description{
get the coordinate reference system (CRS) for simple feature object
}
\examples{
library(sf)
library(spData) # package of spatial data, see also 'spDataLarge'
data(alaska)
data(nz)
data(urban_agglomerations)
see_sf()
get_crs()
}
