% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inla.R
\name{get_inla}
\alias{get_inla}
\title{get character vector of INLA objects in memory}
\usage{
get_inla()
}
\value{
a character vector of INLA objects in memory
}
\description{
get character vector of INLA objects in memory
}
\examples{
get_inla()                                   # nothing if you have no INLA objects
mtcars <- mtcars |> as.data.frame()          # get data
m1 <- INLA::inla(mpg ~ 1   , data = mtcars)  # fit an intercept-only model
m2 <- INLA::inla(mpg ~ disp, data = mtcars)  # fit another model
get_inla()                                   # character vector of models
get_inla() |> mget()                         # list, ripe for map(...)
}
\seealso{
\code{inla_plot_fe} for plotting fixed effects and, \code{inla_table_fe} for extracting tables of fixed effects from INLA objects
}
