% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_plot_fe.R
\name{inla_plot_fe}
\alias{inla_plot_fe}
\title{Plot fixed effects from model fit with INLA::inla}
\usage{
inla_plot_fe(inla_model)
}
\arguments{
\item{inla_model}{}
}
\value{
a ggplot graph (or table with "inla_table_fe") of fixed effects derived from a model fit with INLA::inla()
}
\description{
Plot fixed effects from model fit with INLA::inla
}
\examples{
mtcars <- mtcars |> as.data.frame()
m1 <- inla(mpg ~ 1   , data = mtcars)
m2 <- inla(mpg ~ disp, data = mtcars)
inla_plot_fe(m1)
inla_plot_fe(m2)
}
