% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_table_fe.R
\name{inla_table_fe}
\alias{inla_table_fe}
\title{Make table of fixed effects from model fit with INLA::inla}
\usage{
inla_table_fe(inla_model)
}
\arguments{
\item{inla_model}{}
}
\value{
a table of fixed effects estimated from a model fit with INLA::inla()
}
\description{
Make table of fixed effects from model fit with INLA::inla
}
\examples{
mtcars <- mtcars |> as.data.frame()
m1 <- inla(mpg ~ 1   , data = mtcars)
m2 <- inla(mpg ~ disp, data = mtcars)
inla_table_fe(m1)
inla_table_fe(m2)
get_inla() |> map(inla_table_fe)
}
