% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_sf.R
\name{see_sf}
\alias{see_sf}
\title{get sf objects in memory as character vector}
\usage{
see_sf()
}
\value{
a character vector
}
\description{
get sf objects in memory as character vector
}
\examples{
# These examples only work when copy/pasted into the console since they refer to your environment
see_sf()                 # nothing
library(sf)
test_sf_object <- nz     # a built in dataset in the sf package
see_sf()                 # character string of sf objects
# TODO link to get_crs
}
\seealso{
\code{\link[=see_spatial]{see_spatial()}}, \code{\link[=see_crs]{see_crs()}}
}
