% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_sf.R
\name{see_sf}
\alias{see_sf}
\title{get sf objects in memory as character vector}
\usage{
see_sf()
}
\value{
a character vector
}
\description{
get sf objects in memory as character vector
}
\examples{
see_sf()                 # nothing
library(sfdep)
test_sf_object <- guerry # a built in dataset in the sfdep package
see_sf()                 # character string of sf objects
see_spatial()            # also finds objects of class sp, sf, or RasterLayer
# TODO link to get_crs
}
\seealso{
\code{see_spatial}
}
