% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_sf.R
\name{see_sf}
\alias{see_sf}
\title{get sf object in memory as a character string}
\usage{
see_sf()
}
\value{
a character vector
}
\description{
get sf object in memory as a character string
}
\examples{
see_sf()
library(sfdep)
test_sf_object <- guerry # a built in dataset in the sfdep package
see_sf()
}
