% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_spatial.R
\name{see_spatial}
\alias{see_spatial}
\title{get spatial (sf, sp, RasterLayer) objects in memory as character vector}
\usage{
see_spatial()
}
\value{
a character vector
}
\description{
get spatial (sf, sp, RasterLayer) objects in memory as character vector
}
