% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_spatial.R
\name{see_spatial}
\alias{see_spatial}
\title{get spatial (sf, sp, RasterLayer) objects in memory as character vector}
\usage{
see_spatial()
}
\value{
a character vector
}
\description{
get spatial (sf, sp, RasterLayer) objects in memory as character vector
}
\examples{
see_spatial()                 # nothing
library(sfdep)
test_sf_object <- guerry # a built in dataset in the sfdep package, an sf object
see_spatial()                 # character string of sf objects
see_spatial()            # also finds objects of class sp, sf, or RasterLayer
# TODO link to get_crs
}
\seealso{
\code{\link[=see_sf]{see_sf()}}
}
